[#41] Refactor : FilterButton, Modal 컴포넌트 분리
GuestButton -> FilterButton으로 네이밍 수정 및 별도 컴포넌트로 분리
